@@ -1,6 +1,6 @@
 {
 	"name": "string-width",
-	"version": "3.1.0",
+	"version": "4.2.2",
 	"description": "Get the visual width of a string - the number of columns required to display it",
 	"license": "MIT",
 	"repository": "sindresorhus/string-width",
@@ -10,19 +10,18 @@
 		"url": "sindresorhus.com"
 	},
 	"engines": {
-		"node": ">=6"
+		"node": ">=8"
 	},
 	"scripts": {
-		"test": "xo && ava"
+		"test": "xo && ava && tsd"
 	},
 	"files": [
-		"index.js"
+		"index.js",
+		"index.d.ts"
 	],
 	"keywords": [
 		"string",
-		"str",
 		"character",
-		"char",
 		"unicode",
 		"width",
 		"visual",
@@ -45,12 +44,13 @@
 		"fixed-width"
 	],
 	"dependencies": {
-		"emoji-regex": "^7.0.1",
-		"is-fullwidth-code-point": "^2.0.0",
-		"strip-ansi": "^5.1.0"
+		"emoji-regex": "^8.0.0",
+		"is-fullwidth-code-point": "^3.0.0",
+		"strip-ansi": "^6.0.0"
 	},
 	"devDependencies": {
-		"ava": "^1.0.1",
-		"xo": "^0.23.0"
+		"ava": "^1.4.1",
+		"tsd": "^0.7.1",
+		"xo": "^0.24.0"
 	}
 }
