@@ -54,7 +54,7 @@ const _abi = [
 ];
 
 const _bytecode =
-  "0x608060405234801561001057600080fd5b5060405161011c38038061011c83398101604081905261002f91610049565b600055600180546001600160a01b03191633179055610062565b60006020828403121561005b57600080fd5b5051919050565b60ac806100706000396000f3fe6080604052348015600f57600080fd5b506004361060325760003560e01c80638c3c9a55146037578063f7a3080614604c575b600080fd5b60005460405190815260200160405180910390f35b605c6057366004605e565b600055565b005b600060208284031215606f57600080fd5b503591905056fea2646970667358221220728f6fe7ccc08ce9d12e175d1cf441ee762bc990b5fe5dae84948d9c7bd4cd5a64736f6c63430008070033";
+  "0x608060405234801561001057600080fd5b5060405161011c38038061011c83398101604081905261002f91610049565b600055600180546001600160a01b03191633179055610062565b60006020828403121561005b57600080fd5b5051919050565b60ac806100706000396000f3fe6080604052348015600f57600080fd5b506004361060325760003560e01c80638c3c9a55146037578063f7a3080614604c575b600080fd5b60005460405190815260200160405180910390f35b605c6057366004605e565b600055565b005b600060208284031215606f57600080fd5b503591905056fea264697066735822122094772f26fe8228cfb5275a5e14709e498151179463bf8ab4dfbfb4b1155feb2c64736f6c63430008070033";
 
 export class Dummy__factory extends ContractFactory {
   constructor(signer?: Signer) {
