@@ -23,7 +23,7 @@ const _abi = [
 ];
 
 const _bytecode =
-  "0x60806040526000805460ff19169055348015601957600080fd5b506086806100286000396000f3fe6080604052348015600f57600080fd5b506004361060285760003560e01c8063158ef93e14602d575b600080fd5b60336047565b604080519115158252519081900360200190f35b60005460ff169056fea26469706673582212203d813bc922a43a4d88dada93bca77c6d911f05db6628a85e73e7d725e91cadde64736f6c634300060c0033";
+  "0x60806040526000805460ff19169055348015601957600080fd5b506086806100286000396000f3fe6080604052348015600f57600080fd5b506004361060285760003560e01c8063158ef93e14602d575b600080fd5b60336047565b604080519115158252519081900360200190f35b60005460ff169056fea26469706673582212200c8cf5e36d201baa6944183902cff8e4f448b859c8567ac66a04a534c90f6ff164736f6c634300060c0033";
 
 export class Initializable__factory extends ContractFactory {
   constructor(signer?: Signer) {
