@@ -55,7 +55,7 @@ const _abi = [
 ];
 
 const _bytecode =
-  "0x608060405234801561001057600080fd5b5060405161056f38038061056f83398101604081905261002f916100fe565b600080546001600160a01b0319166001600160a01b038416179055805161005d906001906020840190610065565b505050610241565b828054610071906101f0565b90600052602060002090601f01602090048101928261009357600085556100d9565b82601f106100ac57805160ff19168380011785556100d9565b828001600101855582156100d9579182015b828111156100d95782518255916020019190600101906100be565b506100e59291506100e9565b5090565b5b808211156100e557600081556001016100ea565b6000806040838503121561011157600080fd5b82516001600160a01b038116811461012857600080fd5b602084810151919350906001600160401b038082111561014757600080fd5b818601915086601f83011261015b57600080fd5b81518181111561016d5761016d61022b565b604051601f8201601f19908116603f011681019083821181831017156101955761019561022b565b8160405282815289868487010111156101ad57600080fd5b600093505b828410156101cf57848401860151818501870152928501926101b2565b828411156101e05760008684830101525b8096505050505050509250929050565b600181811c9082168061020457607f821691505b6020821081141561022557634e487b7160e01b600052602260045260246000fd5b50919050565b634e487b7160e01b600052604160045260246000fd5b61031f806102506000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c806306fdde031461003b5780638c3c9a5514610059575b600080fd5b61004361006f565b6040516100509190610222565b60405180910390f35b6100616100fd565b604051908152602001610050565b6001805461007c90610295565b80601f01602080910402602001604051908101604052809291908181526020018280546100a890610295565b80156100f55780601f106100ca576101008083540402835291602001916100f5565b820191906000526020600020905b8154815290600101906020018083116100d857829003601f168201915b505050505081565b60008060008060008060008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663feaf968c6040518163ffffffff1660e01b815260040160a06040518083038186803b15801561016c57600080fd5b505afa158015610180573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906101a491906101d2565b50919998505050505050505050565b805169ffffffffffffffffffff811681146101cd57600080fd5b919050565b600080600080600060a086880312156101ea57600080fd5b6101f3866101b3565b9450602086015193506040860151925060608601519150610216608087016101b3565b90509295509295909350565b600060208083528351808285015260005b8181101561024f57858101830151858201604001528201610233565b81811115610261576000604083870101525b50601f017fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe016929092016040019392505050565b600181811c908216806102a957607f821691505b602082108114156102e3577f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b5091905056fea2646970667358221220119a577f38b91a59ab328c79d20ea0637317a67d86b3a85c4bd31842d0ae72f064736f6c63430008070033";
+  "0x608060405234801561001057600080fd5b5060405161056f38038061056f83398101604081905261002f916100fe565b600080546001600160a01b0319166001600160a01b038416179055805161005d906001906020840190610065565b505050610241565b828054610071906101f0565b90600052602060002090601f01602090048101928261009357600085556100d9565b82601f106100ac57805160ff19168380011785556100d9565b828001600101855582156100d9579182015b828111156100d95782518255916020019190600101906100be565b506100e59291506100e9565b5090565b5b808211156100e557600081556001016100ea565b6000806040838503121561011157600080fd5b82516001600160a01b038116811461012857600080fd5b602084810151919350906001600160401b038082111561014757600080fd5b818601915086601f83011261015b57600080fd5b81518181111561016d5761016d61022b565b604051601f8201601f19908116603f011681019083821181831017156101955761019561022b565b8160405282815289868487010111156101ad57600080fd5b600093505b828410156101cf57848401860151818501870152928501926101b2565b828411156101e05760008684830101525b8096505050505050509250929050565b600181811c9082168061020457607f821691505b6020821081141561022557634e487b7160e01b600052602260045260246000fd5b50919050565b634e487b7160e01b600052604160045260246000fd5b61031f806102506000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c806306fdde031461003b5780638c3c9a5514610059575b600080fd5b61004361006f565b6040516100509190610222565b60405180910390f35b6100616100fd565b604051908152602001610050565b6001805461007c90610295565b80601f01602080910402602001604051908101604052809291908181526020018280546100a890610295565b80156100f55780601f106100ca576101008083540402835291602001916100f5565b820191906000526020600020905b8154815290600101906020018083116100d857829003601f168201915b505050505081565b60008060008060008060008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663feaf968c6040518163ffffffff1660e01b815260040160a06040518083038186803b15801561016c57600080fd5b505afa158015610180573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906101a491906101d2565b50919998505050505050505050565b805169ffffffffffffffffffff811681146101cd57600080fd5b919050565b600080600080600060a086880312156101ea57600080fd5b6101f3866101b3565b9450602086015193506040860151925060608601519150610216608087016101b3565b90509295509295909350565b600060208083528351808285015260005b8181101561024f57858101830151858201604001528201610233565b81811115610261576000604083870101525b50601f017fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe016929092016040019392505050565b600181811c908216806102a957607f821691505b602082108114156102e3577f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b5091905056fea2646970667358221220aef672974ee4fc6b17d0fe0c74ca6b00a75878bb3506794cd102f75179cba98164736f6c63430008070033";
 
 export class PriceConsumerV3__factory extends ContractFactory {
   constructor(signer?: Signer) {
