@@ -21,7 +21,7 @@ const _abi = [
 ];
 
 const _bytecode =
-  "0x608060405234801561001057600080fd5b5060405161015a38038061015a83398101604081905261002f916100c4565b600180546001600160a01b0319166001600160a01b03831690811790915560408051637ff79e0f60e01b81529051637ff79e0f91600480820192602092909190829003018186803b15801561008357600080fd5b505afa158015610097573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906100bb91906100f4565b6000555061010d565b6000602082840312156100d657600080fd5b81516001600160a01b03811681146100ed57600080fd5b9392505050565b60006020828403121561010657600080fd5b5051919050565b603f8061011b6000396000f3fe6080604052600080fdfea264697066735822122007edcce6a37cfc325ac4317739c0eb56964eafedbccc9266f8004781a9649b5964736f6c63430008070033";
+  "0x608060405234801561001057600080fd5b5060405161015a38038061015a83398101604081905261002f916100c4565b600180546001600160a01b0319166001600160a01b03831690811790915560408051637ff79e0f60e01b81529051637ff79e0f91600480820192602092909190829003018186803b15801561008357600080fd5b505afa158015610097573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906100bb91906100f4565b6000555061010d565b6000602082840312156100d657600080fd5b81516001600160a01b03811681146100ed57600080fd5b9392505050565b60006020828403121561010657600080fd5b5051919050565b603f8061011b6000396000f3fe6080604052600080fdfea264697066735822122010df1bd323465537fb7fdec2840933f881c1f0d4429da008bd2ad49280f888a664736f6c63430008070033";
 
 export class VaultRouter__factory extends ContractFactory {
   constructor(signer?: Signer) {
